Comentarios y logs en el servicio UserReviewServiceImpl

+++ api/src/main/java/com/hosting/rest/api/services/User/UserReview/UserReviewServiceImpl.java
@@ -14,9 +14,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.hosting.rest.api.exceptions.IllegalArguments.IllegalArgumentsCustomException;
+import com.hosting.rest.api.exceptions.NotFound.NotFoundCustomException;
 import com.hosting.rest.api.models.User.HostReviewModel;
+import com.hosting.rest.api.repositories.User.IUserRepository;
 import com.hosting.rest.api.repositories.User.UserReview.IUserReviewRepository;
 
+import lombok.extern.slf4j.Slf4j;
+
+/**
+ * 
+ * @author Francisco Coya
+ * @version v1.0.0
+ * @apiNote Servicio que gestiona las valoraciones de un usuario de la
+ *          aplicación.
+ *
+ */
+@Slf4j
 @Service
 public class UserReviewServiceImpl implements IUserReviewService {
 
@@ -26,34 +39,59 @@ public class UserReviewServiceImpl implements IUserReviewService {
 	@Autowired
 	private IUserReviewRepository userReviewRepo;
 
+	@Autowired
+	private IUserRepository userRepo;
+
+	/**
+	 * Añade una nueva valoración a un usuario.
+	 * 
+	 * @param newUserReview
+	 * 
+	 * @return
+	 */
 	@Override
 	public HostReviewModel addNewUserReview(final HostReviewModel newUserReview) {
 		if (!isNotNull(newUserReview)) {
-			throw new IllegalArgumentsCustomException(
-					"Los datos introducidos para el nuevo usuario no son válidos.");
+			log.error("Los datos introducidos para el nuevo usuario no son válidos.");
+			throw new IllegalArgumentsCustomException("Los datos introducidos para el nuevo usuario no son válidos.");
 		}
 
 		boolean existsUserReview = userReviewRepo.findById(newUserReview.getId()).get() != null;
 
 		if (existsUserReview) {
+			log.error("Ya existe una valoración para el usuario.");
 			throw new IllegalArgumentsCustomException("Ya existe una valoración para el usuario.");
 		}
 
 		return userReviewRepo.save(newUserReview);
 	}
 
-	@Override
+	/**
+	 * Actualiza el contenido de una valoración de un usuario.
+	 * 
+	 * @param userId
+	 * @param userReviewToUpdate
+	 * 
+	 * @return
+	 */
 	@Transactional
+	@Override
 	public HostReviewModel updateUserReview(final Integer userId, final HostReviewModel userReviewToUpdate) {
 
 		if (!isIntegerValidAndPositive(userId)) {
-			//throw new IllegalAccomodationArgumentsException("El id del usuario introducido no es válido.");
+			log.error("El id del usuario [ " + userId + " ] introducido no es válido.");
+			throw new IllegalArgumentsCustomException("El id del usuario [ " + userId + " ] introducido no es válido.");
 		}
 
 		HostReviewModel originalHostReview = userReviewRepo.findById(userId).get();
 
 		if (!isNotNull(originalHostReview)) {
-			//throw new AccomodationNotFoundException("La valoración de un usuario a actualizar no existe.");
+			log.error("La valoración a actualizar los datos no existe.");
+			throw new NotFoundCustomException("La valoración a actualizar los datos no existe.");
+		}
+
+		if (!userRepo.existsById(userId)) {
+			throw new NotFoundCustomException("El usuario con id [ " + userId + " ] no existe.");
 		}
 
 		// Contenido de la review
@@ -65,26 +103,42 @@ public class UserReviewServiceImpl implements IUserReviewService {
 		return userReviewRepo.save(originalHostReview);
 	}
 
+	/**
+	 * Elimina una valoración de un usuario por el id de usuario <code>userId</code>
+	 * 
+	 * @param userId
+	 * 
+	 */
 	@Override
 	public void deleteUserReview(final Integer userId) {
 		// TODO: Comprobar id valido
-		
-		// TODO: Comprobar que existe el usuario a borrar
-		
-		userReviewRepo.deleteById(userId);
-	}
-
-	@Override
-	public List<HostReviewModel> findByUserId(final Integer userId) {
-
-		// TODO: Revisar
 		if (!isIntegerValidAndPositive(userId)) {
+			log.error("El id de usuario [ " + userId + " ] no es válido.");
 			throw new IllegalArgumentsCustomException("El id de usuario [ " + userId + " ] no es válido.");
 		}
 
-		/**
-		 * Listado de las valoraciones a un usuario <code>userId</code>.
-		 */
+		if (!userRepo.existsById(userId)) {
+			throw new NotFoundCustomException("El usuario con id [ " + userId + " ] a borrar su valoración no existe.");
+		}
+
+		userReviewRepo.deleteById(userId);
+	}
+
+	/**
+	 * Listado de las valoraciones a un usuario <code>userId</code>.
+	 * 
+	 * @param userId
+	 * 
+	 * @return
+	 */
+	@Override
+	public List<HostReviewModel> findByUserId(final Integer userId) {
+
+		if (!isIntegerValidAndPositive(userId)) {
+			log.error("El id de usuario [ " + userId + " ] no es válido.");
+			throw new IllegalArgumentsCustomException("El id de usuario [ " + userId + " ] no es válido.");
+		}
+
 		String findByUserIdQuery = "SELECT hr " + " FROM HostReviewModel hr " + " INNER JOIN hr.idUserA hu "
 				+ " WHERE hu.id = :userId";
